Going to adjust some comments after merge.

* add comments class in selector package

* actor comments

* 1. comments of collector config
2. index creator support options: overwrite, ignore, off

* fix index testcase bug

* fix comments

* change the config of index creat to index initialize model

* test case failure

+++ skywalking-collector/skywalking-collector-worker/src/main/java/com/a/eye/skywalking/collector/worker/storage/IndexCreator.java
@@ -1,5 +1,6 @@
 package com.a.eye.skywalking.collector.worker.storage;
 
+import com.a.eye.skywalking.collector.worker.config.EsConfig;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -16,10 +17,19 @@ public enum IndexCreator {
     private Logger logger = LogManager.getFormatterLogger(IndexCreator.class);
 
     public void create() {
-        Set<AbstractIndex> indexSet = loadIndex();
-        for (AbstractIndex index : indexSet) {
-            index.deleteIndex();
-            index.createIndex();
+        if (!EsConfig.IndexInitMode.manual.equals(EsConfig.Es.Index.Initialize.mode)) {
+            Set<AbstractIndex> indexSet = loadIndex();
+            for (AbstractIndex index : indexSet) {
+                boolean isExists = index.isExists();
+                if (isExists) {
+                    if (EsConfig.IndexInitMode.forced.equals(EsConfig.Es.Index.Initialize.mode)) {
+                        index.deleteIndex();
+                        index.createIndex();
+                    }
+                } else {
+                    index.createIndex();
+                }
+            }
         }
     }
